Add files via upload
An online coffee shop, also known as an e-commerce coffee shop or virtual coffee shop, is a platform that allows customers to browse and purchase coffee and related products online. Instead of visiting a physical store, customers can access the coffee shop's website or mobile app to explore various coffee options, place orders, and have them delivered to their doorstep. It is just a layout shown in my project.